refactor: use string representation of address in msgs/queries
When we refactored the blacklist to work with base64 encoded bytes vs. the bech32 encoded address we changed some fields in msgs and queries to be a `[]byte`. This creates an unorthodox api for users since most people are accustomed to passing around addresses as strings. To mitigate any confusion we use bech32 encoded addresses in all queries and msgs and do the decoding and conversion inside of the msg handlers.